pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aletheia-ai"
version = "0.1.0"
description = "Aletheia AI - Autonomous Thinking Assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-genai>=1.12.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "pyautogui>=0.9.54",
  "selenium>=4.21.0",
  "Pillow>=10.3.0",
  "opencv-python>=4.10.0.84",
  "numpy>=1.26.0"
]

[project.scripts]
aletheia-ai = "aletheia_ai.main:main"

[tool.setuptools.packages.find]
where = ["src"]
